requirements.txt
discord.py==2.0.1
aioping==0.3.1
aiohttp~=3.8.3
